# internal/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "sprinklerGo API",
    "description": "REST-API der sprinklerGo-Bewässerungssteuerung (Go-Port von sprinklers_pi). Zonen- und Programm-IDs sind 0-basierte Indizes; Startzeiten sind Minuten seit lokaler Mitternacht; Laufzeiten sind Minuten (0-255).",
    "version": "1.0.0",
    "license": { "name": "GPL-2.0", "identifier": "GPL-2.0-only" }
  },
  "paths": {
...
    "/api/auth": {
      "get": {
        "summary": "Auth-Status (immer erreichbar)",
        "operationId": "getAuth",
        "responses": {
          "200": {
            "description": "Status; Token-Liste nur für angemeldete Aufrufer",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "enabled": { "type": "boolean" },
                    "loggedIn": { "type": "boolean" },
                    "hasPassword": { "type": "boolean" },
                    "tokens": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "name": { "type": "string" },
                          "createdAt": { "type": "integer" }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Anmeldung aktivieren/deaktivieren (Passwort muss gesetzt sein)",
        "operationId": "putAuth",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["enabled"],
                "properties": { "enabled": { "type": "boolean" } }
              }
            }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/Ok" },
          "400": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "summary": "Anmelden — setzt ein Session-Cookie (30 Tage)",
        "operationId": "login",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["password"],
                "properties": { "password": { "type": "string" } }
              }
            }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/Ok" },
          "401": { "$ref": "#/components/responses/Error" }
        }
      }
    },
...
    "/api/auth/tokens/{name}": {
      "delete": {
        "summary": "API-Token widerrufen",
        "operationId": "deleteToken",
        "parameters": [
          { "name": "name", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "$ref": "#/components/responses/Ok" },
          "404": { "$ref": "#/components/responses/Error" }
        }
      }
    },
...
  },
  "components": {
...
    "responses": {
      "Ok": {
        "description": "Erfolg",
        "content": {
          "application/json": {
            "schema": { "type": "object", "properties": { "ok": { "type": "boolean" } } }
          }
        }
      },
      "Error": {
        "description": "Fehler",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": ["error"],
              "properties": { "error": { "type": "string" } }
            }
          }
        }
      }
    },
...
  }
}